+++ NEWS.md
@@ -1,8 +1,8 @@
-# jakR2 0.2.4
-
 # jakR2 0.2
 
 * Helpful functions to create and work with palettes and colors
+	* new `order` parameter for default, reverse and random palette order
+	* `blend_palette()` to use monochromeR::generate_palette() blending 
 * Some `jak_theme()` tweaks
 * Alpha-diversity metrics
 * Session and package info utilities